% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coffee_break.R
\name{coffee.break}
\alias{coffee.break}
\title{Take a coffe break}
\usage{
coffee.break(duration = 30)
}
\arguments{
\item{duration}{minutes; choose number of minutes the break should last (greater than 5); default 30}
}
\description{
For the program to work the Power & Sleep settings has to be adjusted.


Please go to:

Settings -> Power & Sleep

There, change each duration to 2 hours.


After running the program keep Microsoft Team as a front window.

Now you are all set!
}
\details{

}
\examples{
coffee.break(duration = 15)
}
